00012-integrate-react-grpc-go-gRPC-server-client Got the Ping / Pong calls to work. Need to cleanup the Makefile
