--- app/cmd/go-swiss-pairing/server.go
@@ -1,1 +1,0 @@
-package main
