+++ app/constants/constants.go
@@ -3,3 +3,4 @@ package constants
 const CertDir = "./cert/"
 const ServerCertPem = CertDir + "server-cert.pem"
 const ServerKeyPem = CertDir + "server-key.pem"
+const ServerCrt = CertDir + "server.crt"
